fix(wit-bindgen-wrpc): clear out internal state after use
This commit fixes a bug which manifests itself when decoding a list of
enums that may have separate internal decoder objects to use.

The issue is that storing the internal state inside the bindgen'd
`Decoder` becomes a problem on the *second* (or subsequent) elements
of a list of the given enum.

Given a `vec![SomeEnum::X, SomeEnum::Y]` the saved "state" for the
first becomes the first decoder (i.e. the decoder for
`SomeEnum::X`). When it comes time to parse the discriminant and bytes
for `Y` the existence of the `state` causes a short circuit.

As the (wrong) decoder will generally try to process some bytes, you
lose bytes from the input and then bytes are left over at the next
attempt to decode (and then an error occurs). How this manifests is
obviously specific to the enum, (wrong) decoder, and involved
variants, how they are decoded -- i.e. where the decoding will fail.

This only happens for generating bindings for variant types (i.e. Rust
enums), which use the `PayloadDecoder` machinery to dynamically
pick *which* payload decoder to use, depending on the discriminant read.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>